ml-service/requirements.txt
# ── ML Service Python Dependencies ────────────────────────────────────────
#
# Install with:
#   pip install -r requirements.txt
#
# Pin versions in production with `pip freeze > requirements-lock.txt`.

# ── Web framework ──────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0        # ASGI server with WebSocket & HTTP/2 support
python-multipart>=0.0.9          # Required by FastAPI for file upload parsing

# ── Config management ──────────────────────────────────────────────────────
pydantic-settings>=2.2.0         # .env → pydantic Settings class
python-dotenv>=1.0.0             # Load .env file automatically

# ── HTTP client (for calling Ollama / OpenAI) ──────────────────────────────
httpx>=0.27.0                    # Async HTTP client used by ML service

# ── Document parsing ───────────────────────────────────────────────────────
pymupdf>=1.24.0                  # PDF parsing (import as `fitz`)
pdfplumber>=0.11.0               # PDF parsing (high-precision/tables)

python-docx>=1.1.0               # DOCX parsing (Open XML / .docx files)

pytesseract>=0.3.10              # OCR bridge (requires Tesseract binary installed)
pdf2image>=1.17.0                # Converts PDF pages to images for OCR
pillow>=10.3.0                   # Image handling for PIL/Pillow

pytesseract>=0.3.10              # OCR bridge (requires Tesseract binary installed)
pdf2image>=1.17.0                # Converts PDF pages to images for OCR
pillow>=10.3.0                   # Image handling for PIL/Pillow

# ── Text chunking ──────────────────────────────────────────────────────────
langchain-text-splitters>=0.0.1  # RecursiveCharacterTextSplitter

# ── Embeddings ─────────────────────────────────────────────────────────────
sentence-transformers>=2.7.0     # Local embedding model (all-MiniLM-L6-v2)
                                 # Runs on CPU; no API key required

# ── Vector store ───────────────────────────────────────────────────────────
chromadb>=0.5.0                  # Local persistent vector database

# ── LLM (optional — only needed if using OpenAI) ──────────────────────────
openai>=1.30.0                   # OpenAI API client (GPT-4, embeddings)
